feat: updated modal comps + react-tilt add
