+++ src/components/space/index.ts
@@ -277,3 +277,15 @@ export default class SpaceContainer {
   // }
 
 }
+
+// if ( WebGL.isWebGLAvailable() ) {
+
+// 	// Initiate function or other initializations here
+// 	animate();
+
+// } else {
+
+// 	const warning = WebGL.getWebGLErrorMessage();
+// 	document.getElementById( 'container' ).appendChild( warning );
+
+// }
