+++ src/index.css
@@ -26,6 +26,7 @@ body,
 #root {
     height: 100%;
     width: 100%;
+    background: var(--baseBlack);
 }
 
 div.close-modal {
